+++ src/components/forms/jungeun/LoginInputField.jsx
@@ -1,6 +1,10 @@
 "use client"
 
+import { useState } from "react"
+
 const InputField = ({ type, placeholder, value, onChange, required, icon, error }) => {
+  const [showPassword, setShowPassword] = useState(false)
+  const [inputType, setInputType] = useState(type)
 
   const focusStyles = {
     borderColor: error ? "#ff6b6b" : "#FDCD00",
@@ -28,6 +32,11 @@ const InputField = ({ type, placeholder, value, onChange, required, icon, error 
     }
   }
 
+  const togglePasswordVisibility = () => {
+    setShowPassword(!showPassword)
+    setInputType(showPassword ? "password" : "text")
+  }
+
   const getIcon = () => {
     switch (icon) {
       case "id":
@@ -47,6 +56,20 @@ const InputField = ({ type, placeholder, value, onChange, required, icon, error 
     }
   }
 
+  const getPasswordToggleIcon = () => {
+    if (type !== "password") return null
+    
+    return showPassword ? (
+      <svg width="20" height="20" viewBox="0 0 24 24" fill="currentColor">
+        <path d="M12 4.5C7 4.5 2.73 7.61 1 12c1.73 4.39 6 7.5 11 7.5s9.27-3.11 11-7.5c-1.73-4.39-6-7.5-11-7.5zM12 17c-2.76 0-5-2.24-5-5s2.24-5 5-5 5 2.24 5 5-2.24 5-5 5zm0-8c-1.66 0-3 1.34-3 3s1.34 3 3 3 3-1.34 3-3-1.34-3-3-3z" />
+      </svg>
+    ) : (
+      <svg width="20" height="20" viewBox="0 0 24 24" fill="currentColor">
+        <path d="M12 7c2.76 0 5 2.24 5 5 0 .65-.13 1.26-.36 1.83l2.92 2.92c1.51-1.26 2.7-2.89 3.43-4.75-1.73-4.39-6-7.5-11-7.5-1.4 0-2.74.25-3.98.7l2.16 2.16C10.74 7.13 11.35 7 12 7zM2 4.27l2.28 2.28.46.46C3.08 8.3 1.78 10.02 1 12c1.73 4.39 6 7.5 11 7.5 1.55 0 3.03-.3 4.38-.84l.42.42L19.73 22 21 20.73 3.27 3 2 4.27zM7.53 9.8l1.55 1.55c-.05.21-.08.43-.08.65 0 1.66 1.34 3 3 3 .22 0 .44-.03.65-.08l1.55 1.55c-.67.33-1.41.53-2.2.53-2.76 0-5-2.24-5-5 0-.79.2-1.53.53-2.2zm4.31-.78l3.15 3.15.02-.16c0-1.66-1.34-3-3-3l-.17.01z" />
+      </svg>
+    )
+  }
+
   return (
     <div className={`input-field-container${error ? ' error' : ''}`}>
       <div className={`input-icon${error ? ' error' : ''}`}>
@@ -54,7 +77,7 @@ const InputField = ({ type, placeholder, value, onChange, required, icon, error 
       </div>
       <input
         className={`input-field${error ? ' error' : ''}`}
-        type={type}
+        type={inputType}
         placeholder={placeholder}
         value={value}
         onChange={onChange}
@@ -62,6 +85,37 @@ const InputField = ({ type, placeholder, value, onChange, required, icon, error 
         onFocus={handleFocus}
         onBlur={handleBlur}
       />
+      {type === "password" && (
+        <button
+          type="button"
+          className="password-toggle-btn"
+          onClick={togglePasswordVisibility}
+          style={{
+            position: "absolute",
+            right: "16px",
+            top: "50%",
+            transform: "translateY(-50%)",
+            background: "none",
+            border: "none",
+            cursor: "pointer",
+            color: "rgba(23, 15, 88, 0.6)",
+            padding: "4px",
+            display: "flex",
+            alignItems: "center",
+            justifyContent: "center",
+            transition: "color 0.3s ease",
+            zIndex: 2
+          }}
+          onMouseEnter={(e) => {
+            e.target.style.color = "#FDCD00"
+          }}
+          onMouseLeave={(e) => {
+            e.target.style.color = "rgba(23, 15, 88, 0.6)"
+          }}
+        >
+          {getPasswordToggleIcon()}
+        </button>
+      )}
       <style>
         {`
           input::placeholder {
