+++ src/components/layout/jungeun/TestHeader.jsx
@@ -1,13 +1,10 @@
 import React from "react";
 import { logout } from "../../../api/Auth";
 import useAuthStore from "../../../store/jungeun/AuthStore";
-import { useToast } from "../../../context/jungeun/ToastContext";
 
 
 const Header = () => {
 
-    const { showToast } = useToast();
-    
     const handleLogout = async (e) => {
         e.preventDefault()
 
@@ -33,18 +30,34 @@ const Header = () => {
             alignItems: "center",
             padding: "10px 20px",
             backgroundColor: "#170F58",
-            color: "white"
+            color: "white",
+            boxShadow: "0 2px 8px rgba(23, 15, 88, 0.15)"
         }}>
-            <h1 style={{ margin: 0 }}>TESSERIS</h1>
+            <h1 style={{ margin: 0, fontWeight: "600" }}>TESSERIS</h1>
             <button
                 onClick={handleLogout}
                 style={{
                     backgroundColor: "#FDCD00",
                     border: "none",
                     padding: "8px 16px",
-                    borderRadius: "4px",
+                    borderRadius: "8px",
                     cursor: "pointer",
-                    fontWeight: "bold"
+                    fontWeight: "600",
+                    color: "#170F58",
+                    transition: "all 0.3s ease",
+                    boxShadow: "0 2px 4px rgba(253, 205, 0, 0.2)",
+                    fontFamily: "-apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, 'Helvetica Neue', Arial, sans-serif",
+                    fontSize: "14px"
+                }}
+                onMouseEnter={(e) => {
+                    e.target.style.backgroundColor = "#FDD835";
+                    e.target.style.transform = "translateY(-1px)";
+                    e.target.style.boxShadow = "0 4px 8px rgba(253, 205, 0, 0.3)";
+                }}
+                onMouseLeave={(e) => {
+                    e.target.style.backgroundColor = "#FDCD00";
+                    e.target.style.transform = "translateY(0)";
+                    e.target.style.boxShadow = "0 2px 4px rgba(253, 205, 0, 0.2)";
                 }}
             >
                 로그아웃
