+++ src/styles/jungeun/confirmModal.css
@@ -1,0 +1,143 @@
+.modal-backdrop {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background: rgba(0, 0, 0, 0.6);
+  backdrop-filter: blur(12px) saturate(180%);
+  -webkit-backdrop-filter: blur(12px) saturate(180%);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  z-index: 1000;
+  animation: fadeIn 0.3s ease-out;
+  will-change: backdrop-filter;
+}
+
+.modal-content {
+  background: rgba(245, 245, 249, 0.95);
+  backdrop-filter: blur(25px) saturate(180%);
+  -webkit-backdrop-filter: blur(25px) saturate(180%);
+  border: 1px solid rgba(253, 205, 0, 0.2);
+  border-radius: 20px;
+  padding: 2.5rem;
+  text-align: center;
+  box-shadow: 
+    0 20px 40px rgba(23, 15, 88, 0.1),
+    0 0 0 1px rgba(253, 205, 0, 0.1),
+    inset 0 1px 0 rgba(255, 255, 255, 0.2);
+  max-width: 400px;
+  width: 90%;
+  position: relative;
+  animation: slideUp 0.4s ease-out;
+  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, sans-serif;
+  will-change: backdrop-filter;
+}
+
+.modal-content::before {
+  content: '';
+  position: absolute;
+  top: -2px;
+  left: -2px;
+  right: -2px;
+  bottom: -2px;
+  background: linear-gradient(135deg, #FDCD00, #FDD835, #170F58);
+  border-radius: 22px;
+  z-index: -1;
+  opacity: 0.1;
+}
+
+.modal-content p {
+  margin: 0 0 2rem 0;
+  font-size: 1.1rem;
+  font-weight: 500;
+  color: #170F58;
+  line-height: 1.5;
+}
+
+.modal-content button {
+  background: linear-gradient(135deg, #FDCD00 0%, #FDD835 100%);
+  color: #170F58;
+  border: none;
+  padding: 12px 32px;
+  border-radius: 12px;
+  font-size: 1rem;
+  font-weight: 600;
+  cursor: pointer;
+  transition: all 0.3s ease;
+  box-shadow: 
+    0 4px 15px rgba(253, 205, 0, 0.3),
+    0 2px 4px rgba(23, 15, 88, 0.1);
+  position: relative;
+  overflow: hidden;
+}
+
+.modal-content button::before {
+  content: '';
+  position: absolute;
+  top: 0;
+  left: -100%;
+  width: 100%;
+  height: 100%;
+  background: linear-gradient(90deg, transparent, rgba(255, 255, 255, 0.4), transparent);
+  transition: left 0.5s ease;
+}
+
+.modal-content button:hover {
+  transform: translateY(-2px);
+  box-shadow: 
+    0 8px 25px rgba(253, 205, 0, 0.4),
+    0 4px 8px rgba(23, 15, 88, 0.15);
+  background: linear-gradient(135deg, #FDD835 0%, #FDCD00 100%);
+}
+
+.modal-content button:hover::before {
+  left: 100%;
+}
+
+.modal-content button:active {
+  transform: translateY(0);
+  box-shadow: 
+    0 2px 8px rgba(253, 205, 0, 0.3),
+    0 1px 2px rgba(23, 15, 88, 0.1);
+}
+
+@keyframes fadeIn {
+  from {
+    opacity: 0;
+  }
+  to {
+    opacity: 1;
+  }
+}
+
+@keyframes slideUp {
+  from {
+    opacity: 0;
+    transform: translateY(30px) scale(0.9);
+  }
+  to {
+    opacity: 1;
+    transform: translateY(0) scale(1);
+  }
+}
+
+/* 반응형 디자인 */
+@media (max-width: 480px) {
+  .modal-content {
+    padding: 2rem 1.5rem;
+    margin: 1rem;
+    border-radius: 16px;
+  }
+  
+  .modal-content p {
+    font-size: 1rem;
+    margin-bottom: 1.5rem;
+  }
+  
+  .modal-content button {
+    padding: 10px 24px;
+    font-size: 0.9rem;
+  }
+}
